TeleBot/requirements.txt
Django==2.0 <2
amqp==2.3.2
asgiref==3.2.3
asn1crypto==0.24.0
astroid==2.3.3
atomicwrites==1.4.0
attrs==19.3.0
azure-cognitiveservices-vision-computervision==0.6.0
azure-common==1.1.25
bcrypt==3.1.4
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
comtypes==1.1.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Cython==0.29.19
decorator==4.4.2
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-configurations==2.1
django-cors-headers==3.2.1
django-dotenv==1.4.2
django-filter==2.2.0
django-phonenumbers==1.0.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-rest-swagger==2.2.0
django-timezone-field==4.0
docopt==0.6.2
docx2pdf==0.1.7
drf-yasg==1.17.0
graphviz==0.13.2
httplib2==0.16.0
idna==2.8
image==1.5.32
imageio==2.8.0
inflection==0.3.1
isodate==0.6.0
isort==4.3.21
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10
joblib==0.15.1
kiwisolver==1.2.0
kombu==4.6.7
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.2.1
mccabe==0.6.1
more-itertools==8.3.0
mrz==0.5.8
msrest==0.6.14
networkx==2.4
num2words==0.5.10
numpy==1.18.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.1
pandas==0.24.2
PassportEye==1.4.1
pdfminer3k==1.3.1
phonenumbers==8.11.2
Pillow==7.0.0
plantuml==0.3.0
pluggy==0.13.1
ply==3.11
psycopg2==2.8.4
psycopg2-binary==2.8.4
py==1.8.1
pycparser==2.18
PyJWT==1.7.1
pylint==2.4.4
PyNaCl==1.3.0
pyparsing==2.4.6
PyPDF2==1.26.0
pytesseract==0.3.4
pytest==5.4.2
python-crontab==2.4.0
python-dateutil==2.8.1
python-dotenv==0.9.1
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.7
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
tesseract==0.1.3
threadpoolctl==2.0.0
tifffile==2020.5.25
tqdm==4.46.0
uritemplate==3.0.0
urllib3==1.25.7
vine==1.1.4
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
wcwidth==0.1.9
wrapt==1.11.2
xlrd==1.1.0
XlsxWriter==1.2.7
xlwt==1.3.0
xmltodict==0.12.0
